cmc: code format

+++ tools/cmc/client/dispatch.go
@@ -50,7 +50,8 @@ func runInvokeContract(client *sdk.ChainClient, contractName, method string, par
 	}()
 
 	for i := 0; i < totalCntPerGoroutine; i++ {
-		resp, err := client.InvokeContract(contractName, method, "", util.ConvertParameters(params), int64(timeout), syncResult)
+		resp, err := client.InvokeContract(contractName, method, "", util.ConvertParameters(params),
+			timeout, syncResult)
 		if err != nil {
 			fmt.Printf("[ERROR] invoke contract failed, %s", err.Error())
 			return
@@ -61,7 +62,8 @@ func runInvokeContract(client *sdk.ChainClient, contractName, method string, par
 			return
 		}
 
-		fmt.Printf("INVOKE contract resp, [code:%d]/[msg:%s]/[contractResult:%+v]\n", resp.Code, resp.Message, resp.ContractResult)
+		fmt.Printf("INVOKE contract resp, [code:%d]/[msg:%s]/[contractResult:%+v]\n", resp.Code, resp.Message,
+			resp.ContractResult)
 	}
 }
 func runInvokeContractOnce(client *sdk.ChainClient, contractName, method string, params map[string]string,
@@ -70,7 +72,8 @@ func runInvokeContractOnce(client *sdk.ChainClient, contractName, method string,
 	defer func() {
 		wg.Done()
 	}()
-	resp, err := client.InvokeContract(contractName, method, txId, util.ConvertParameters(params), int64(timeout), syncResult)
+	resp, err := client.InvokeContract(contractName, method, txId, util.ConvertParameters(params), int64(timeout),
+		syncResult)
 	if err != nil {
 		fmt.Printf("[ERROR] invoke contract failed, %s", err.Error())
 		return
@@ -81,6 +84,6 @@ func runInvokeContractOnce(client *sdk.ChainClient, contractName, method string,
 		return
 	}
 
-	fmt.Printf("INVOKE contract resp, [code:%d]/[msg:%s]/[contractResult:%+v]\n", resp.Code, resp.Message, resp.ContractResult)
-
+	fmt.Printf("INVOKE contract resp, [code:%d]/[msg:%s]/[contractResult:%+v]\n", resp.Code, resp.Message,
+		resp.ContractResult)
 }
